Fix hash field types and remove TOKEN auth, add mandatory BLAKE3 hashing
This commit makes three major improvements to the VSF format:

1. **Fixed hash field types to preserve algorithm information**
   - Changed hash fields from `Option<Vec<u8>>` to `Option<(u8, Vec<u8>)>`
   - Affects: dark_frame_hash, flat_field_hash, bias_frame_hash,
     vignette_correction_hash, distortion_correction_hash
   - Updated in RawMetadata struct, RawMetadataBuilder, and parser
   - Encoding now uses stored algorithm instead of hardcoded HASH_BLAKE3
   - Parsing now extracts and stores both algorithm byte and hash bytes
   - All tests updated to use tuple format: `(HASH_BLAKE3, vec![0xAB; 32])`

2. **Removed TOKEN authentication system**
   - Removed TokenAuth and TokenBuilder structs completely
   - Removed token field from RawImageBuilder
   - Removed token_auth parameter from build_raw_image()
   - Removed all TOKEN parsing logic from parse_raw_image()
   - Removed token_auth field from ParsedRawImage struct
   - Updated all examples and tests to remove TOKEN usage
   - Removed TOKEN-related documentation and comments
   - Rationale: VSF already has hash (h) and signature (g) types for
     verification - TOKEN was redundant and a "weird tangent"

3. **Made BLAKE3 file hashing automatic and mandatory**
   - Every VSF file now automatically includes a BLAKE3 hash in the header
   - Changed VsfBuilder::new() to default include_file_hash: true
   - Removed with_file_hash() method (no longer needed - always enabled)
   - VsfBuilder::build() now automatically calls add_file_hash()
   - Parser updated to skip file hash field after version/backward compat
   - Updated tests to reflect automatic hashing behavior
   - New test: test_verify_hash_integrity() to verify corruption detection
   - Documentation updated to reflect that hashing is transparent and required

Breaking changes:
- Hash fields now require tuple format: (algorithm, bytes)
- TOKEN authentication no longer available (use h/g types instead)
- All VSF files now include file hash (adds ~40 bytes to header)

All 94 tests passing.

+++ examples/create_raw_image.rs
@@ -28,7 +28,7 @@ fn main() -> Result<(), String> {
     for sample in 0..total_samples {
         rng ^= rng.rotate_left(13).wrapping_add(sample);
         let value = rng as u8 as u16 + blackpoint; // Simulate RAW image data
-        samples.push(value as u64);
+        samples.push(value);
     }
 
     // Pack into BitPackedTensor (efficient storage for 11-bit samples)
@@ -54,12 +54,12 @@ fn main() -> Result<(), String> {
     ]);
 
     // Set camera settings
-    raw.camera.iso_speed = Some(800.0);
-    raw.camera.shutter_time_s = Some(1.0 / 60.0); // 1/60 second
+    raw.camera.iso_speed = Some(800.);
+    raw.camera.shutter_time_s = Some(1. / 60.); // 1/60 second
     raw.camera.aperture_f_number = Some(2.8);
     raw.camera.focal_length_m = Some(0.024); // 24mm = 0.024m
-    raw.camera.exposure_compensation = Some(0.0);
-    raw.camera.focus_distance_m = Some(5.0);
+    raw.camera.exposure_compensation = Some(0.);
+    raw.camera.focus_distance_m = Some(5.);
     raw.camera.flash_fired = Some(false);
     raw.camera.metering_mode = Some("matrix".to_string());
 
@@ -70,7 +70,7 @@ fn main() -> Result<(), String> {
     raw.lens.min_focal_length_m = Some(0.024); // Prime lens
     raw.lens.max_focal_length_m = Some(0.024);
     raw.lens.min_aperture_f = Some(2.8);
-    raw.lens.max_aperture_f = Some(22.0);
+    raw.lens.max_aperture_f = Some(22.);
 
     // Build the VSF file
     let raw_bytes = raw.build()?;
